Add the Mapper
Actually, there's only an SQLMapper for now.  It's supposed to be the
DataMapper layer or whatever.  The whole MVC thing is still new to me
and I have not understood it yet.  The fact that everyone on the web
seems to have their own idea of what it is does not help either.  In any
case, from what I got, this seems to be a step in the right direction.

This does not fix the problem though, that the Treeview saves everything
as strings and so when I try to do the update later, Postgres rightfully
complains that '7' is not the correct representation for an integer
field.