Remove controller.py
Gone for good.  It's not really too clear to me what the lifetime of a
controller is supposed to be.  So, the obvious sollution is to just
remove them altogether.  Their jobs will just be done by the views.
It's probably not clean, but hey, for a project this size it's not a
problem.